requirements.txt
tensorflow
tensorflow_hub
numpy
python-dotenv
matplotlib
aiogram
pydantic